#6724 : A48 - Ecran de saisie d'une demande depuis SIRH

+++ src/main/java/nc/mairie/gestionagent/absence/dto/CompteurAsaDto.java
@@ -6,7 +6,7 @@ public class CompteurAsaDto {
 
 	private Integer idAgent;
 
-	private Integer nb;
+	private Double nb;
 
 	private Date dateDebut;
 
@@ -36,11 +36,11 @@ public class CompteurAsaDto {
 		this.dateFin = dateFin;
 	}
 
-	public Integer getNb() {
+	public Double getNb() {
 		return nb;
 	}
 
-	public void setNb(Integer nb) {
+	public void setNb(Double nb) {
 		this.nb = nb;
 	}
